+++ website/theme/components/Landingpage/Hero/index.module.scss
