+++ website/theme/components/Landingpage/Hero/index.tsx
@@ -1,6 +1,6 @@
 import { useCallback } from 'react';
 import { useNavigate } from 'rspress/runtime';
-import { useI18n, useI18nUrl } from '../../../theme/i18n';
+import { useI18n, useI18nUrl } from '../../../i18n';
 import styles from './index.module.scss';
 
 const Hero = () => {
